rename loss functions
