+++ nugraph/nugraph/util/__init__.py
@@ -1,8 +1,8 @@
 """Loss functions, data transforms and general utilities"""
 from .confusion_matrix_logger import ConfusionMatrixLogger
 from .recall_loss import RecallLoss
-from .LogCoshLoss import LogCoshLoss
-from .ObjCondensationLoss import ObjCondensationLoss
+from .log_cosh_loss import LogCoshLoss
+from .obj_con_loss import ObjConLoss
 from .position_features import PositionFeatures
 from .feature_extension import FeatureExtension
 from .hierarchical_edges import HierarchicalEdges
