+++ nugraph/nugraph/util/log_cosh_loss.py
